fix Sam addresses on linux as well
It turns out the fixes I had to do since I switched to darwin were not
OS-related. It simply was that the Sam addresses are working differently
now compared to when I initially wrote xplor. Probably something changed
in the p9p implementation of them. So the fixes I did for darwin are
actually needed now for linux (and others) as well.

Fixes #5